Crate det

det is a command line tool for email related DNS records.

You can use det to check the following DNS records:

Sender Policy Framework (SPF)

Usage

Check the SPF record for a domain

det spf example.com

you can provide a specific SPF record to check

det spf example.com --record "v=spf1 -all"

you can also get more details about the SPF record

det spf example.com --detail

Modules

common

Common module

dns

DNS module

spf

SPF module

Structs

Cli 🔒

Enums

Commands 🔒

Functions

main 🔒